ENH: remove pixel detector base, add to spot detector, move spots up one level
