--- starfish/pipeline/features/pixels/_base.py
@@ -1,7 +1,0 @@
-from starfish.pipeline.algorithmbase import AlgorithmBase
-
-
-class PixelFinderAlgorithmBase(AlgorithmBase):
-    def find(self, stack):
-        """Find spots."""
-        raise NotImplementedError()
