+++ starfish/pipeline/spots/_validated_table.py
