+++ starfish/pipeline/spots/decoder/_base.py
