+++ starfish/pipeline/spots/detector/_base.py
@@ -6,7 +6,7 @@ import xarray as xr
 from starfish._stack import ImageStack
 from starfish.intensity_table import IntensityTable
 from starfish.pipeline.algorithmbase import AlgorithmBase
-from starfish.pipeline.features.spot_attributes import SpotAttributes
+from starfish.pipeline.spots.spot_attributes import SpotAttributes
 from starfish.types import Number
