+++ starfish/pipeline/spots/detector/combine_adjacent_features.py
