+++ starfish/pipeline/spots/spot_attributes.py
