+++ starfish/test/full_pipelines/api/test_iss_api.py
@@ -1,4 +1,4 @@
-from starfish.pipeline.features.spots.detector.gaussian import GaussianSpotDetector
+from starfish.pipeline.spots.detector.gaussian import GaussianSpotDetector
 from starfish.image._filter.white_tophat import WhiteTophat
 from starfish.image._registration.fourier_shift import FourierShiftRegistration
 from starfish.util.synthesize import SyntheticData
